Update demo 3
