--- Teams/04 Fundamentals of Microsoft Teams/Demos/03 - teams-app1/src/app/web/tou.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="UTF-8">
-  <title>
-    teams app1 - Terms of Use
-  </title>
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <!-- inject:css -->
-  <!-- endinject -->
-</head>
-
-<body>
-  <h1>
-    teams app1 Terms of Use</h1>
-  <p>
-    Add your Terms of Use here...
-  </p>
-</body>
-
-</html>
