+++ Teams/04 Fundamentals of Microsoft Teams/Demos/03-tab-graph/solution/.yo-rc.json
@@ -1,0 +1,7 @@
+{
+  "generator-teams": {
+    "promptValues": {
+      "developer": "Office Developer"
+    }
+  }
+}
