+++ Teams/04 Fundamentals of Microsoft Teams/Demos/03-tab-graph/solution/package.json
@@ -13,34 +13,35 @@
     "@types/express": "4.16.0",
     "@types/express-session": "1.15.10",
     "@types/morgan": "1.7.35",
+    "@types/react": "16.4.7",
     "@types/request": "2.47.1",
     "botbuilder": "3.15.0",
     "botbuilder-teams": "0.2.1",
     "browser-request": "0.3.3",
     "debug": "^3.1.0",
+    "dotenv": "6.0.0",
     "express": "4.16.3",
     "express-msteams-host": "~1.0.0",
     "express-session": "1.15.6",
-    "dotenv": "6.0.0",
     "fancy-log": "1.3.2",
+    "file-loader": "1.1.11",
     "gulp": "4.0.0",
     "gulp-inject": "4.3.2",
     "gulp-zip": "4.2.0",
     "morgan": "1.9.0",
+    "msal": "^0.2.3",
+    "msteams-react-base-component": "~1.0.0",
+    "msteams-ui-components-react": "^0.7.3",
     "nodemon": "1.18.3",
     "plugin-error": "1.0.1",
+    "react": "^16.1.0",
+    "react-dom": "^16.2.0",
     "request": "2.87.0",
     "ts-loader": "4.4.2",
     "typescript": "2.6.1",
+    "typestyle": "1.5.1",
     "webpack": "4.16.3",
     "yargs": "12.0.1",
-    "z-schema": "^3.19.1",
-    "msteams-react-base-component": "~1.0.0",
-    "msteams-ui-components-react": "^0.7.3",
-    "react": "^16.1.0",
-    "@types/react": "16.4.7",
-    "react-dom": "^16.2.0",
-    "file-loader": "1.1.11",
-    "typestyle": "1.5.1"
+    "z-schema": "^3.19.1"
   }
 }
