+++ Teams/04 Fundamentals of Microsoft Teams/Demos/03-tab-graph/solution/src/app/web/index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="UTF-8">
+    <title>
+        teams app1
+    </title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <!-- inject:css -->
+    <!-- endinject -->
+</head>
+
+<body>
+    <h1>
+        Welcome to the <em>teams app1</em></h1>
+    <p>
+        Generate a Microsoft Teams application.
+    </p>
+    
+</body>
+
+</html>
