+++ Teams/04 Fundamentals of Microsoft Teams/Demos/03-tab-graph/solution/src/app/web/teamsApp1TabConfig.html
@@ -10,14 +10,14 @@
 </head>
 
 <body>
-  <div id='app'>
-    Loading...
-  </div>
-  <!-- inject:js -->
-  <!-- endinject -->
-  <script type='text/javascript'>
-    teamsApp1.teamsApp1TabConfig.render(document.getElementById('app'), {});
-  </script>
+    <div id='app'>
+        Loading...
+      </div>
+      <!-- inject:js -->
+      <!-- endinject -->
+      <script type='text/javascript'>
+        teamsApp1.teamsApp1TabConfig.render(document.getElementById('app'), {});
+      </script>
 
 </body>
