+++ Teams/04 Fundamentals of Microsoft Teams/Demos/03-tab-graph/solution/src/manifest/icon-outline.png
