+++ Teams/04 Fundamentals of Microsoft Teams/Demos/03-tab-graph/solution/src/manifest/manifest.json
@@ -1,14 +1,14 @@
 {
-  "$schema": "https://developer.microsoft.com/en-us/json-schemas/teams/v1.3/MicrosoftTeams.schema.json",
-  "manifestVersion": "1.3",
-  "id": "[app-id-from-registration]",
+  "$schema": "https://statics.teams.microsoft.com/sdk/v1.2/manifest/MicrosoftTeams.schema.json",
+  "manifestVersion": "1.2",
+  "id": "d0bfd1a4-aab0-72f6-0bb8-1400d18f3cb5",
   "version": "0.0.1",
-  "packageName": "io.ngrok.tbd",
+  "packageName": "io.ngrok.dbef3d7d",
   "developer": {
     "name": "Office Developer",
-    "websiteUrl": "https://tbd.ngrok.io",
-    "privacyUrl": "https://tbd.ngrok.io/privacy.html",
-    "termsOfUseUrl": "https://tbd.ngrok.io/tou.html"
+    "websiteUrl": "https://dbef3d7d.ngrok.io",
+    "privacyUrl": "https://dbef3d7d.ngrok.io/privacy.html",
+    "termsOfUseUrl": "https://dbef3d7d.ngrok.io/tou.html"
   },
   "name": {
     "short": "teams app1",
@@ -25,7 +25,7 @@
   "accentColor": "#D85028",
   "configurableTabs": [
     {
-      "configurationUrl": "https://tbd.ngrok.io/teamsApp1TabConfig.html",
+      "configurationUrl": "https://dbef3d7d.ngrok.io/teamsApp1TabConfig.html",
       "canUpdateConfiguration": true,
       "scopes": [
         "team"
@@ -41,6 +41,6 @@
     "messageTeamMembers"
   ],
   "validDomains": [
-    "tbd.ngrok.io"
+    "dbef3d7d.ngrok.io"
   ]
 }
